products/Private-AI-Lab/apps/private-doc-intel-demo/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-multipart
pypdf
python-docx
fastembed
